libm : fmod : better fmod and fmodf

+++ libm/generic/fmodf.c
@@ -1,9 +1,9 @@
 #include <math.h>
 #include <limits.h>
 
-double fmod(double x, double y) {
+float fmodf(float x, float y) {
 	if (isnan(x) || isnan(y) || isinf(x) || y == 0) return NAN;
 	if (isinf(y)) return x;
-	double quotien = trunc(x / y);
+	float quotien = truncf(x / y);
 	return x - quotien * y;
 }
